Refactor user functions test (#5873)
* Refactor graphQL and concurrency functions api tests

* moved the config and queries to separate files

* Added missing licences

* Merged with refactor_graphql and tranferred functions from common_test file
